# backend/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Contrôle Parental API'
  description: "API REST pour l'application de Contrôle Parental - gestion des familles, enfants, appareils, règles de filtrage, temps d'écran, alertes et rapports."
  version: 1.0.0
servers:
  - url: 'http://localhost:8000'
    description: 'Serveur local'
paths:
  /api/activities:
    get:
      tags:
        - Activities
      summary: 'Liste des activités'
      description: 'Retourne les activités (paginées), filtrable par child_id et device_id'
      operationId: activitiesIndex
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
        -
          name: child_id
          in: query
          schema:
            type: integer
        -
          name: device_id
          in: query
          schema:
            type: integer
      responses:
        200:
          description: 'Liste paginée des activités'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Liste des activités'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  meta:
                    type: object
                  links:
                    type: object
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          sanctum: []
  '/api/activities/{activity}':
    get:
      tags:
        - Activities
      summary: "Détails d'une activité"
      operationId: activitiesShow
      parameters:
        -
          name: activity
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: "Détails de l'activité"
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Détails de l'activité"
                  data:
                    $ref: '#/components/schemas/Activity'
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
...
  /api/devices:
    get:
      tags:
        - Devices
      summary: 'Liste des appareils'
      description: 'Retourne les appareils (paginés), filtrable par child_id'
      operationId: devicesIndex
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
        -
          name: child_id
          in: query
          schema:
            type: integer
      responses:
        200:
          description: 'Liste paginée des appareils'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Liste des appareils'
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
                  meta:
                    type: object
                  links:
                    type: object
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          sanctum: []
    post:
      tags:
        - Devices
      summary: 'Ajouter un appareil'
      operationId: devicesStore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - child_id
                - name
                - type
              properties:
                child_id:
                  type: integer
                  example: 1
                name:
                  type: string
                  example: 'iPhone de Lucas'
                type:
                  type: string
                  example: smartphone
                os:
                  type: string
                  example: iOS
                os_version:
                  type: string
                  example: '17.2'
              type: object
      responses:
        201:
          description: 'Appareil ajouté'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Appareil ajouté'
                  data:
                    $ref: '#/components/schemas/Device'
                type: object
        422:
          $ref: '#/components/responses/ValidationError'
      security:
        -
          sanctum: []
  '/api/devices/{device}':
    get:
      tags:
        - Devices
      summary: "Détails d'un appareil"
      operationId: devicesShow
      parameters:
        -
          name: device
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: "Détails de l'appareil"
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Détails de l'appareil"
                  data:
                    $ref: '#/components/schemas/Device'
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
    put:
      tags:
        - Devices
      summary: 'Mettre à jour un appareil'
      operationId: devicesUpdate
      parameters:
        -
          name: device
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: 'iPhone de Lucas'
                status:
                  type: string
                  example: paired
                  enum:
                    - paired
                    - unpaired
                    - locked
                    - lost
                is_online:
                  type: boolean
                  example: true
                battery_level:
                  type: integer
                  example: 75
              type: object
      responses:
        200:
          description: 'Appareil mis à jour'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Appareil mis à jour'
                  data:
                    $ref: '#/components/schemas/Device'
                type: object
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Devices
      summary: 'Supprimer un appareil'
      operationId: devicesDestroy
      parameters:
        -
          name: device
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
      security:
        -
          sanctum: []
...
components:
  schemas:
...
    Device:
      properties:
        id:
          type: integer
          example: 1
        child_id:
          type: integer
          example: 1
        name:
          type: string
          example: 'iPhone de Lucas'
        type:
          type: string
          example: smartphone
        os:
          type: string
          example: iOS
        os_version:
          type: string
          example: '17.2'
        app_version:
          type: string
          nullable: true
        status:
          type: string
          example: paired
        is_online:
          type: boolean
          example: true
        battery_level:
          type: integer
          example: 75
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        paired_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    Activity:
      properties:
        id:
          type: integer
          example: 1
        child_id:
          type: integer
          example: 1
        device_id:
          type: integer
          example: 1
        type:
          type: string
          example: app_usage
        target:
          type: string
          example: com.youtube.app
        duration_seconds:
          type: integer
          example: 3600
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
      type: object
...
  responses:
    Success:
      description: 'Opération réussie'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: true
              message:
                type: string
                example: 'Opération réussie'
              data:
                type: object
                nullable: true
            type: object
    Unauthorized:
      description: 'Non authentifié'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Unauthenticated.
            type: object
    ValidationError:
      description: 'Erreur de validation'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 'Validation error'
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
            type: object
    NotFound:
      description: 'Ressource introuvable'
      content:
        application/json:
          schema:
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 'Ressource introuvable'
            type: object
...
  securitySchemes:
    sanctum:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header
...
security:
  -
    sanctum: []